backend/requirements.txt
fastapi
uvicorn
sqlalchemy
bcrypt
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]
pydantic-settings
email-validator
python-dotenv
